+++ fbpcs/CHANGELOG.md
@@ -15,6 +15,7 @@ Types of changes
 
 ## [Unreleased - 2.6.0] - put release date here
 ### Removed
+* Removed deprecated PCSMPCInstance/MPCInstance from fbpcs
 * Removed deprecated MPCInstanceRepository from fbpcs
 * Removed deprecated MPCService APIs after MPCInstance usaged been moved to StageStateInstance. (create/start/stop/get/update_instance)
 * Removed `get_mpc` from private_computation_cli endpoint